+++ apps/client/src/app/modules/commission-board/user-rating-details-popup/user-rating-details-popup.component.ts
@@ -2,6 +2,7 @@ import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { CommissionProfile } from '../../../model/user/commission-profile';
 import { NzModalRef } from 'ng-zorro-antd/modal';
 import { TranslateService } from '@ngx-translate/core';
+import { SettingsService } from '../../settings/settings.service';
 
 @Component({
   selector: 'app-user-rating-details-popup',
@@ -12,7 +13,7 @@ import { TranslateService } from '@ngx-translate/core';
 export class UserRatingDetailsPopupComponent {
   profile: CommissionProfile;
 
-  constructor(private modalRef: NzModalRef, public translate: TranslateService) {
+  constructor(private modalRef: NzModalRef, public translate: TranslateService, public settings: SettingsService) {
   }
 
   close(): void {